去除sql语句where 1=1

+++ trunk/DTCMS.BLL/Arc_ClassBLL.cs
@@ -187,7 +187,7 @@ namespace DTCMS.BLL
 		/// </summary>
         /// <param name="count">返回记录数</param>
         /// <returns>对象泛型集合</returns>
-        public List<Arc_Class> GetList(out long count)
+        public List<Arc_Class> GetList(out int count)
         {
             return dal.GetList(out count);
         }
@@ -199,7 +199,7 @@ namespace DTCMS.BLL
         /// <param name="pageIndex">当前页</param>
         /// <param name="count">返回记录数</param>
         /// <returns>分页对象泛型集合</returns>
-        public List<Arc_Class> GetPageList(int pageSize, int pageIndex, out long count)
+        public List<Arc_Class> GetPageList(int pageSize, int pageIndex, out int count)
         {
             if (pageSize <= 0)
                 throw new Exception("每页数据条数必须大于0。");
