+++ trunk/DTCMS.BLL/UsersBLL.cs
@@ -80,7 +80,7 @@ namespace DTCMS.BLL
 		/// </summary>
         /// <param name="count">返回记录数</param>
         /// <returns>对象泛型集合</returns>
-        public List<Users> GetList(out long count)
+        public List<Users> GetList(out int count)
         {
             return dal.GetList(out count);
         }
@@ -92,7 +92,7 @@ namespace DTCMS.BLL
         /// <param name="pageIndex">当前页</param>
         /// <param name="count">返回记录数</param>
         /// <returns>分页对象泛型集合</returns>
-        public List<Users> GetPageList(int pageSize, int pageIndex, out long count)
+        public List<Users> GetPageList(int pageSize, int pageIndex, out int count)
         {
             if (pageSize <= 0)
                 throw new Exception("每页数据条数必须大于0。");
